backend/requirements.txt
numpy==1.26
Pillow
uvicorn
fastapi
huggingface_hub
sam2